+++ src/app/components/requeststatus/show/requeststatus-show/requeststatus-show.component.ts
@@ -1,8 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { Requeststate } from 'src/app/model/requeststatus';
-import { Router, ActivatedRoute, Params} from '@angular/router';
+import { Router, ActivatedRoute, Params } from '@angular/router';
 import { RequeststateService } from 'src/app/services/requeststatus.services'
 import { Machinestate } from 'src/app/model/machinestate';
+import swal from 'sweetalert';
 @Component({
   selector: 'app-requeststatus-show',
   templateUrl: './requeststatus-show.component.html',
@@ -15,24 +16,41 @@ export class RequeststatusShowComponent implements OnInit {
   constructor(
     private _requeststateService: RequeststateService,
     private _route: ActivatedRoute,
-    private _router: Router 
+    private _router: Router
   ) {
     this.title = 'Request Status'
-    
-   }
+
+  }
 
   ngOnInit(): void {
-    this._requeststateService.getRequestStates().subscribe(response=>{
-      this.requeststatus=response;
+    this._requeststateService.getRequestStates().subscribe(response => {
+      this.requeststatus = response;
     })
   }
 
-  deleterequeststate(code){
-    this._requeststateService.deleteRequeststate(code).subscribe(
-      response =>{
-        this.ngOnInit();
-      }
-    )
+  deleterequeststate(code) {
+    swal({
+      title: "Estas seguro que deseas eliminarlo?",
+      text: "cuidado estas apundo de eliminar un registro",
+      icon: "warning",
+      buttons: [true, true],
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          swal("Poof! su registro ha sido eliminado", {
+            icon: "success",
+          });
+          this._requeststateService.deleteRequeststate(code).subscribe(
+            response => {
+              this.ngOnInit();
+            }
+          )
+        } else {
+          swal("Tranquilo/a su registro no se ha borrado");
+        }
+      });
+
   }
 
 }
