+++ src/app/components/statebranchoffice/show/statebranchoffice-show/statebranchoffice-show.component.ts
@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { Router, ActivatedRoute, Params} from '@angular/router';
+import { Router, ActivatedRoute, Params } from '@angular/router';
 import { StateBranchOffice } from 'src/app/model/stateBranchOffice';
 import { StateBranchOfficeService } from 'src/app/services/statebranchoffice.services';
+import swal from 'sweetalert';
 @Component({
   selector: 'app-statebranchoffice-show',
   templateUrl: './statebranchoffice-show.component.html',
@@ -14,23 +15,39 @@ export class StatebranchofficeShowComponent implements OnInit {
   constructor(
     private _stateBranchOfficeService: StateBranchOfficeService,
     private _route: ActivatedRoute,
-    private _router: Router 
-  ) { 
+    private _router: Router
+  ) {
     this.title = 'Estado Sucursal'
   }
 
   ngOnInit(): void {
-    this._stateBranchOfficeService.getStatesBranchOffice().subscribe(response=>{
-      this.statebranchOffice=response;
+    this._stateBranchOfficeService.getStatesBranchOffice().subscribe(response => {
+      this.statebranchOffice = response;
     })
   }
 
-  deletestatebranchoffice(code){
-    this._stateBranchOfficeService.deleteStateBranchOffice(code).subscribe(
-      response =>{
-        this.ngOnInit();
-      }
-    )
+  deletestatebranchoffice(code) {
+    swal({
+      title: "Estas seguro que deseas eliminarlo?",
+      text: "cuidado estas apundo de eliminar un registro",
+      icon: "warning",
+      buttons: [true, true],
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          swal("Poof! su registro ha sido eliminado", {
+            icon: "success",
+          });
+          this._stateBranchOfficeService.deleteStateBranchOffice(code).subscribe(
+            response => {
+              this.ngOnInit();
+            }
+          )
+        } else {
+          swal("Tranquilo/a su registro no se ha borrado");
+        }
+      });
+    }
+  
   }
-
-}
